+++ visualization/conf/webpack.loaders.js
@@ -14,6 +14,10 @@ module.exports = {
             use: ['style-loader', 'css-loader']
         },
         {
+            test: /\.scss$/,
+            use: ['style-loader', 'css-loader', 'sass-loader']
+        },
+        {
             test: /\.(png|svg|jpg|gif)$/, use: ['file-loader']
         },
         {