+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ### Fixed
 - Delta View shows Deltas of itself as non-trivial if nodes have same name #89: Compare deltas by path not name
+- Delta calculation performance boost #91
+- Problems when intermediate nodes missed metrics #92
+- removed unnecessary calculations
 
 ## [1.9.3] - 2018-02-23
 ### Added