+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 - Problems when intermediate nodes missed metrics #92
 - removed unnecessary calculations
 - removed bug in SonarImporter that slowed up performance and missed out multiple metrics
+- less flickering and artifacts
 
 ## [1.9.3] - 2018-02-23
 ### Added